shell/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
# --- Autentykacja OIDC + JWT ---
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
itsdangerous>=2.1.2
# --- Testy ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
